[WIP] Working at new login screen. Login components extracted.
